[BE] style: TechStack, UsingTechStack 분리
- service & repository 분리

- issue (#21)
