+++ server/src/domains/techstack/repository/UsingTechStackRepository.ts
@@ -8,4 +8,14 @@ export class UsingTechStackRepository extends Repository<UsingTechStack> {
   public findAll() {
     return this.find();
   }
+
+  public async findUsingTechStackListByGroupId(groupId: number): Promise<any[]> {
+    const groupTechStackJoinResult = await this.createQueryBuilder('using_tech_stack')
+      .innerJoinAndSelect('using_tech_stack.techStack', 'tech_stack')
+      .where('using_tech_stack.type = :type', { type: 'GROUP' })
+      .andWhere('using_tech_stack.groupId = :groupId', { groupId })
+      .getMany();
+
+    return groupTechStackJoinResult;
+  }
 }
