+++ server/src/domains/techstack/service/TechStackService.ts
@@ -1,8 +1,6 @@
 import { Service } from 'typedi';
 import { TechStackRepository } from '../repository/TechStackRepository';
 import { InjectRepository } from 'typeorm-typedi-extensions';
-import { destructObject } from '@utils/Object';
-import { GroupTechStackListDto } from '../dto/groupTechStackListDto';
 
 @Service()
 export class TechStackService {
@@ -14,15 +12,5 @@ export class TechStackService {
   public async getTechStackList() {
     const techStackList = await this.techStackRepository.findAll();
     return techStackList;
-  }
-
-  public async getGroupsTechStackList(groupId: number) {
-    const techStackList = await this.techStackRepository.findTechStackListByGroupId(groupId);
-    return Promise.all(
-      techStackList.map((techStack) => {
-        const { name } = destructObject(techStack) as GroupTechStackListDto;
-        return name;
-      }),
-    );
   }
 }
